+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- Truncate tool call outputs automatically to avoid hit context size limit. #284
+  - Save output to eca cache folde and tell LLM where the full content is.
+  - Add `toolCall outputTruncation` config to customize size or lines.
+
 ## 0.98.5
 
 - Add openai via subscription `gpt-5.3-codex` model.